requirements.txt
numpy==1.23.5
openpyxl==3.0.10
openpyxl==3.1.2
pandas==1.5.2
pyjanitor==0.26.0
PyYAML==6.0.1
PyYAML==6.0.1
streamlit==1.27.2
streamlit_authenticator==0.2.3
